handling when there are no categories or editors created

+++ app/views/admin/viewcategories.view.php
@@ -22,18 +22,23 @@
     <!-- VIEW EDITORS HEADER -->
     <h1 class="viewcategories-header">CATEGORIES</h1>
 
-    <!-- EDITORS GRID -->
-    <div class="category-grid">
-        <?php foreach ($data["categories"] as $category) { ?>
-            <div class="category-card">
-                <div class="category-card__info">
-                    <span class="category-card__name"><?php echo $category["name"]; ?></span>
+    <?php if (count($data["categories"]) == 0) : ?>
+        <span style="font-size: 1.25rem; margin: 1rem; text-align: center;">There are no categories!</span>
+    <?php endif ?>
+
+    <?php if (count($data["categories"]) > 0) : ?>
+        <div class="category-grid">
+            <?php foreach ($data["categories"] as $category) : ?>
+                <div class="category-card">
+                    <div class="category-card__info">
+                        <span class="category-card__name"><?php echo $category["name"]; ?></span>
+                    </div>
+                    <div class="category-card__actions">
+                        <a class="btn btn--white" href="/admin/editcategory?id=<?php echo $category["id"]; ?>">EDIT CATEGORY</a>
+                        <a class="delete-btn btn btn--black" href="/admin/deletecategory?id=<?php echo $category["id"]; ?>">DELETE CATEGORY</a>
+                    </div>
                 </div>
-                <div class="category-card__actions">
-                    <a class="btn btn--white" href="/admin/editcategory?id=<?php echo $category["id"]; ?>">EDIT CATEGORY</a>
-                    <a class="delete-btn btn btn--black" href="/admin/deletecategory?id=<?php echo $category["id"]; ?>">DELETE CATEGORY</a>
-                </div>
-            </div>
-        <?php } ?>
-    </div>
+            <?php endforeach ?>
+        </div>
+    <?php endif ?>
 </div>
